chore: delete test-utils, rename client-primites to node
