+++ bin/node/Cargo.toml
@@ -15,7 +15,7 @@ workspace = true
 [dependencies]
 # internal
 base-cli-utils.workspace = true
-base-client-primitives.workspace = true
+base-client-node.workspace = true
 base-flashblocks.workspace = true
 base-metering.workspace = true
 base-txpool.workspace = true
