--- crates/client/client-primitives/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "base-client-primitives"
-description = "Primitive types and traits for Base node runner extensions"
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-homepage.workspace = true
-repository.workspace = true
-
-[lints]
-workspace = true
-
-[dependencies]
-# reth
-reth.workspace = true
-reth-db.workspace = true
-reth-optimism-node.workspace = true
-reth-optimism-chainspec.workspace = true
-
-# misc
-eyre.workspace = true
-futures-util.workspace = true
-tracing.workspace = true
-derive_more = { workspace = true, features = ["debug"] }
