+++ crates/client/flashblocks/benches/pending_state.rs
@@ -8,11 +8,11 @@ use std::{
 use alloy_eips::{BlockHashOrNumber, Encodable2718};
 use alloy_primitives::{Address, B256, BlockNumber, Bytes, U256, bytes, hex::FromHex};
 use alloy_rpc_types_engine::PayloadId;
+use base_client_node::test_utils::{Account, LocalNodeProvider, TestHarness};
 use base_flashblocks::{FlashblocksAPI, FlashblocksReceiver, FlashblocksState};
 use base_flashtypes::{
     ExecutionPayloadBaseV1, ExecutionPayloadFlashblockDeltaV1, Flashblock, Metadata,
 };
-use base_test_utils::{Account, LocalNodeProvider, TestHarness};
 use criterion::{BatchSize, Criterion, Throughput, criterion_group, criterion_main};
 use reth::{
     chainspec::{ChainSpecProvider, EthChainSpec},
