+++ crates/client/flashblocks/tests/eip7702_tests.rs
@@ -8,11 +8,13 @@ use alloy_eips::{eip2718::Encodable2718, eip7702::Authorization};
 use alloy_primitives::{Address, B256, Bytes, U256};
 use alloy_provider::Provider;
 use alloy_sol_types::SolCall;
+use base_client_node::test_utils::{
+    Account, L1_BLOCK_INFO_DEPOSIT_TX, Minimal7702Account, SignerSync,
+};
 use base_flashblocks::test_harness::FlashblocksHarness;
 use base_flashtypes::{
     ExecutionPayloadBaseV1, ExecutionPayloadFlashblockDeltaV1, Flashblock, Metadata,
 };
-use base_test_utils::{Account, L1_BLOCK_INFO_DEPOSIT_TX, Minimal7702Account, SignerSync};
 use eyre::Result;
 use op_alloy_network::ReceiptResponse;
