+++ crates/client/metering/tests/meter_rpc.rs
@@ -4,8 +4,8 @@ use alloy_eips::Encodable2718;
 use alloy_primitives::{Bytes, U256, address, bytes};
 use alloy_rpc_client::RpcClient;
 use base_bundles::{Bundle, MeterBundleResponse};
+use base_client_node::test_utils::{Account, TestHarness};
 use base_metering::MeteringExtension;
-use base_test_utils::{Account, TestHarness};
 use op_alloy_consensus::OpTxEnvelope;
 use reth_optimism_primitives::OpTransactionSigned;
 use reth_transaction_pool::test_utils::TransactionBuilder;
