+++ crates/client/node/Cargo.toml
@@ -1,0 +1,94 @@
+[package]
+name = "base-client-node"
+description = "Primitive types and traits for Base node runner extensions"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[lints]
+workspace = true
+
+[features]
+test-utils = [
+	"alloy-eips",
+	"alloy-genesis",
+	"alloy-primitives",
+	"alloy-provider",
+	"alloy-rpc-client",
+	"alloy-rpc-types",
+	"alloy-rpc-types-engine",
+	"alloy-signer",
+	"base-primitives/test-utils",
+	"chrono",
+	"jsonrpsee",
+	"op-alloy-network",
+	"op-alloy-rpc-types-engine",
+	"reth-db/test-utils",
+	"reth-ipc",
+	"reth-node-core",
+	"reth-optimism-rpc",
+	"reth-primitives-traits",
+	"reth-provider",
+	"reth-rpc-layer",
+	"reth-tracing",
+	"tokio",
+	"tower",
+	"tracing-subscriber",
+	"url",
+	"reth-optimism-node/test-utils",
+	"reth-primitives-traits?/test-utils",
+	"reth-provider?/test-utils"
+]
+
+[dependencies]
+# Project
+base-primitives = { workspace = true, optional = true }
+
+# reth
+reth.workspace = true
+reth-db.workspace = true
+reth-optimism-node.workspace = true
+reth-optimism-chainspec.workspace = true
+reth-ipc = { workspace = true, optional = true }
+reth-tracing = { workspace = true, optional = true }
+reth-rpc-layer = { workspace = true, optional = true }
+reth-provider = { workspace = true, optional = true }
+reth-node-core = { workspace = true, optional = true }
+reth-primitives-traits = { workspace = true, optional = true }
+reth-optimism-rpc = { workspace = true, features = ["client"], optional = true }
+reth-optimism-primitives.workspace = true
+
+# alloy
+alloy-primitives = { workspace = true, optional = true }
+alloy-genesis = { workspace = true, optional = true }
+alloy-eips = { workspace = true, optional = true }
+alloy-rpc-types = { workspace = true, optional = true }
+alloy-rpc-types-engine = { workspace = true, optional = true }
+alloy-provider = { workspace = true, optional = true }
+alloy-rpc-client = { workspace = true, optional = true }
+alloy-signer = { workspace = true, optional = true }
+
+# op-alloy
+op-alloy-rpc-types-engine = { workspace = true, optional = true }
+op-alloy-network = { workspace = true, optional = true }
+
+# tokio
+tokio = { workspace = true, optional = true }
+
+# rpc
+jsonrpsee = { workspace = true, optional = true }
+
+# misc
+eyre.workspace = true
+futures-util.workspace = true
+tracing.workspace = true
+derive_more = { workspace = true, features = ["debug"] }
+tracing-subscriber = { workspace = true, optional = true }
+url = { workspace = true, optional = true }
+chrono = { workspace = true, optional = true }
+
+# tower for middleware
+tower = { version = "0.5", optional = true }
