+++ crates/client/node/src/extension.rs
