+++ crates/client/node/src/lib.rs
@@ -3,6 +3,9 @@
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 #![cfg_attr(not(test), warn(unused_crate_dependencies))]
 
+// Silence false positive - reth_optimism_primitives is used in test_utils module
+use reth_optimism_primitives as _;
+
 mod extension;
 pub use extension::BaseNodeExtension;
 
@@ -14,3 +17,6 @@ pub use runner::BaseNodeRunner;
 
 mod types;
 pub use types::{BaseNodeBuilder, OpBuilder, OpProvider};
+
+#[cfg(feature = "test-utils")]
+pub mod test_utils;
