--- crates/client/test-utils/Cargo.toml
@@ -1,62 +1,0 @@
-[package]
-name = "base-test-utils"
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-homepage.workspace = true
-repository.workspace = true
-description = "Common integration test utilities for node-reth crates"
-
-[lints]
-workspace = true
-
-[dependencies]
-# Project
-base-client-primitives.workspace = true
-base-primitives = { workspace = true, features = ["test-utils"] }
-
-# reth
-reth.workspace = true
-reth-optimism-node.workspace = true
-reth-optimism-chainspec.workspace = true
-reth-optimism-primitives.workspace = true
-reth-optimism-rpc = { workspace = true, features = ["client"] }
-reth-provider.workspace = true
-reth-primitives-traits.workspace = true
-reth-db = { workspace = true, features = ["test-utils"] }
-reth-node-core.workspace = true
-reth-tracing.workspace = true
-reth-rpc-layer.workspace = true
-reth-ipc.workspace = true
-
-# alloy
-alloy-primitives.workspace = true
-alloy-genesis.workspace = true
-alloy-eips.workspace = true
-alloy-rpc-types.workspace = true
-alloy-rpc-types-engine.workspace = true
-alloy-provider.workspace = true
-alloy-rpc-client.workspace = true
-alloy-signer.workspace = true
-
-# op-alloy
-op-alloy-rpc-types-engine.workspace = true
-op-alloy-network.workspace = true
-
-# tokio
-tokio.workspace = true
-
-# rpc
-jsonrpsee.workspace = true
-
-# misc
-tracing-subscriber.workspace = true
-eyre.workspace = true
-url.workspace = true
-chrono.workspace = true
-
-# tower for middleware
-tower = "0.5"
-
-[dev-dependencies]
