+++ crates/client/txpool/Cargo.toml
@@ -13,7 +13,7 @@ workspace = true
 
 [dependencies]
 # workspace
-base-client-primitives.workspace = true
+base-client-node.workspace = true
 
 # reth
 reth.workspace = true
